--- teacherClass.rb
@@ -1,12 +1,0 @@
-require './person'
-
-Class Teacher < Person
-  def initialize(specialization)
-    super
-    @specialization = specialization
-  end
-
-  def can_use_services?
-    true
-  end
-end
